// src/app/interfaces/anuncio-interface.ts
// Generated by https://quicktype.io

export interface ResultAnuncios {
    ok:      boolean;
    msg:     string;
    anuncio: Anuncio[];
}
export interface ResultAnuncio {
    ok:      boolean;
    msg:     string;
    anuncio: Anuncio;
}

export interface Anuncio {
    _id:         string;
    descripcion: string;
    estado:      boolean;
    usuario:     Usuario;
    fecha:       string;
}

export interface Usuario {
    _id:    string;
    nombre: string;
}